requirements.txt
pyqt6
autobahn
pycryptodome
des